extract History, Activity Handler
