+++ gitit2.cabal
@@ -51,6 +51,7 @@ library
                      Network.Gitit2.Page
 
     other-modules: Network.Gitit2.Cache
+                   Network.Gitit2.Handler.History
                    Network.Gitit2.Handler.View
                    Network.Gitit2.Helper
                    Network.Gitit2.Import
